--- SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/allocator/Machines.java
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2021 The University of Manchester
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package uk.ac.manchester.spinnaker.allocator;
-
-import java.util.List;
-
-class Machines {
-	/** The machine info. */
-	List<BriefMachineDescription> machines;
-
-	public void setMachines(List<BriefMachineDescription> machines) {
-		this.machines = machines;
-	}
-}
